IPRT/posix/semwait.h: Build fix. bugref:10138
svn:sync-xref-src-repo-rev: r148699

+++ src/VBox/Runtime/r3/posix/semwait.h
@@ -1,4 +1,4 @@
-/* $Id: semwait.h 92780 2021-12-07 10:26:50Z knut.osmundsen@oracle.com $ */
+/* $Id: semwait.h 92781 2021-12-07 10:31:02Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - Common semaphore wait code.
  */
@@ -119,7 +119,7 @@ DECLINLINE(uint64_t) rtSemPosixCalcDeadline(uint32_t fFlags, uint64_t uTimeout, 
 
         pAbsDeadline->tv_sec  += TsAdd.tv_sec;
         pAbsDeadline->tv_nsec += TsAdd.tv_nsec;
-        if (pAbsDeadline->tv_nsec >= RT_NS_1SEC)
+        if ((uint32_t)pAbsDeadline->tv_nsec >= RT_NS_1SEC)
         {
             pAbsDeadline->tv_nsec -= RT_NS_1SEC;
             pAbsDeadline->tv_sec++;
